+++ guidata/widgets/objecteditor.py
@@ -16,18 +16,12 @@ This package provides a generic object editor widget.
 
 
 import PIL.Image
-from numpy.core.multiarray import ndarray
-
-from qtpy.QtCore import QObject
 from guidata.widgets.arrayeditor import ArrayEditor
 from guidata.widgets.collectionseditor import CollectionsEditor
-from guidata.widgets.nsview import (
-    DataFrame,
-    FakeObject,
-    Series,
-    is_known_type,
-)
+from guidata.widgets.nsview import DataFrame, FakeObject, Series, is_known_type
 from guidata.widgets.texteditor import TextEditor
+from numpy.core.multiarray import ndarray
+from qtpy.QtCore import QObject
 
 try:
     from guidata.widgets.dataframeeditor import DataFrameEditor
@@ -35,56 +29,7 @@ except ImportError:
     DataFrameEditor = FakeObject()
 
 
-class DialogKeeper(QObject):
-    """ """
-
-    def __init__(self):
-        QObject.__init__(self)
-        self.dialogs = {}
-        self.namespace = None
-
-    def set_namespace(self, namespace):
-        """
-
-        :param namespace:
-        """
-        self.namespace = namespace
-
-    def create_dialog(self, dialog, refname, func):
-        """
-
-        :param dialog:
-        :param refname:
-        :param func:
-        """
-        self.dialogs[id(dialog)] = dialog, refname, func
-        dialog.accepted.connect(lambda eid=id(dialog): self.editor_accepted(eid))
-        dialog.rejected.connect(lambda eid=id(dialog): self.editor_rejected(eid))
-        dialog.show()
-        dialog.activateWindow()
-        dialog.raise_()
-
-    def editor_accepted(self, dialog_id):
-        """
-
-        :param dialog_id:
-        """
-        dialog, refname, func = self.dialogs[dialog_id]
-        self.namespace[refname] = func(dialog)
-        self.dialogs.pop(dialog_id)
-
-    def editor_rejected(self, dialog_id):
-        """
-
-        :param dialog_id:
-        """
-        self.dialogs.pop(dialog_id)
-
-
-keeper = DialogKeeper()
-
-
-def create_dialog(obj, obj_name):
+def create_dialog(obj, title, parent=None):
     """Creates the editor dialog and returns a tuple (dialog, func) where func
     is the function to be called with the dialog instance as argument, after
     quitting the dialog box
@@ -97,26 +42,26 @@ def create_dialog(obj, obj_name):
     conv_func = lambda data: data
     readonly = not is_known_type(obj)
     if isinstance(obj, ndarray):
-        dialog = ArrayEditor()
-        if not dialog.setup_and_check(obj, title=obj_name, readonly=readonly):
+        dialog = ArrayEditor(parent)
+        if not dialog.setup_and_check(obj, title=title, readonly=readonly):
             return
     elif isinstance(obj, PIL.Image.Image):
-        dialog = ArrayEditor()
+        dialog = ArrayEditor(parent)
         import numpy as np
 
         data = np.array(obj)
-        if not dialog.setup_and_check(data, title=obj_name, readonly=readonly):
+        if not dialog.setup_and_check(data, title=title, readonly=readonly):
             return
         conv_func = lambda data: PIL.Image.fromarray(data, mode=obj.mode)
     elif isinstance(obj, (DataFrame, Series)) and DataFrame is not FakeObject:
-        dialog = DataFrameEditor()
+        dialog = DataFrameEditor(parent)
         if not dialog.setup_and_check(obj):
             return
     elif isinstance(obj, str):
-        dialog = TextEditor(obj, title=obj_name, readonly=readonly)
+        dialog = TextEditor(obj, title=title, readonly=readonly, parent=parent)
     else:
-        dialog = CollectionsEditor()
-        dialog.setup(obj, title=obj_name, readonly=readonly)
+        dialog = CollectionsEditor(parent)
+        dialog.setup(obj, title=title, readonly=readonly)
 
     def end_func(dialog):
         """
@@ -129,7 +74,7 @@ def create_dialog(obj, obj_name):
     return dialog, end_func
 
 
-def oedit(obj, modal=True, namespace=None):
+def oedit(obj, title=None, parent=None):
     """Edit the object 'obj' in a GUI-based editor and return the edited copy
     (if Cancel is pressed, return None)
 
@@ -137,38 +82,12 @@ def oedit(obj, modal=True, namespace=None):
 
     Supported container types:
     dict, list, tuple, str/unicode or numpy.array
-
-    (instantiate a new QApplication if necessary,
-    so it can be called directly from the interpreter)
     """
-
-    from guidata import qapplication
-
-    app = qapplication()
-
-    if modal:
-        obj_name = ""
-    else:
-        assert isinstance(obj, str)
-        obj_name = obj
-        if namespace is None:
-            namespace = globals()
-        keeper.set_namespace(namespace)
-        obj = namespace[obj_name]
-        # keep QApplication reference alive in the Python interpreter:
-        namespace["__qapp__"] = app
-
-    result = create_dialog(obj, obj_name)
+    title = "" if title is None else title
+    result = create_dialog(obj, title, parent)
     if result is None:
         return
     dialog, end_func = result
-
-    if modal:
-        if dialog.exec_():
-            return end_func(dialog)
-    else:
-        keeper.create_dialog(dialog, obj_name, end_func)
-        import os
-
-        if os.name == "nt":
-            app.exec_()
+    if dialog.exec_():
+        return end_func(dialog)
+    return None